Previous fix only did delay the issue
My previous fix attempt only made it clear that the issue was
in the runtime and an upstream issue with Webpack, but did not
really fix things.

Since Webpack 5.96.0, especially since
https://github.com/webpack/webpack/commit/420d0d0eed79b0c300f18466a47ed7bd5f4ff14b
Webpack does not generate JS for asset chunks, which breaks having
a single entry with both JS and asset chunks.

The logo can easily be moved to a separate entry.